+++ .github/workflows/build.yml
@@ -9,7 +9,7 @@ jobs:
     - uses: actions/checkout@v3
     - uses: actions/setup-python@v4
       with:
-        python-version: "3.10"
+        python-version: "3.11"
         cache: "pip"
     - run: pip install -r requirements.txt
     - run: pytest
