+++ .github/workflows/scripts.yml
@@ -4,7 +4,7 @@ on:
     branches:
       - master
   schedule:
-    - cron: "0 */72 * * *"  # every three days
+    - cron: "0 0 */3 * *"  # every three days
 jobs:
   build:
     runs-on: ubuntu-latest
@@ -12,7 +12,7 @@ jobs:
     - uses: actions/checkout@v3
     - uses: actions/setup-python@v4
       with:
-        python-version: "3.10"
+        python-version: "3.11"
         cache: "pip"
     - run: pip install -r requirements.txt
     - run: python ./scripts/inbox.py
