--- Procfile
@@ -1,1 +1,0 @@
-web: gunicorn film2trello:app
