+++ fly.toml
@@ -1,35 +1,5 @@
-# fly.toml app configuration file generated for film2trello on 2023-05-23T21:54:25+02:00
-#
-# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
-#
-
 app = "film2trello"
 primary_region = "fra"
 kill_signal = "SIGINT"
 kill_timeout = "5s"
-
-[build]
-  builtin = "python"
-
-[[services]]
-  protocol = "tcp"
-  internal_port = 8080
-  processes = ["app"]
-
-  [[services.ports]]
-    port = 80
-    handlers = ["http"]
-
-  [[services.ports]]
-    port = 443
-    handlers = ["tls", "http"]
-  [services.concurrency]
-    type = "connections"
-    hard_limit = 25
-    soft_limit = 20
-
-  [[services.tcp_checks]]
-    interval = "15s"
-    timeout = "2s"
-    grace_period = "10s"
-    restart_limit = 0
+processes = []
